weather_pipeline/requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
apache-airflow==2.9.1
